src/main/webapp/app/core/home/home.component.ts
import { type ComputedRef, defineComponent, inject } from 'vue';
import { useStore } from '@/store';

import type LoginService from '@/account/login.service';

export default defineComponent({
  compatConfig: { MODE: 3 },
  setup() {
    const loginService = inject<LoginService>('loginService');

    const authenticated = inject<ComputedRef<boolean>>('authenticated');
    const username = inject<ComputedRef<string>>('currentUsername');

    const store = useStore();

    const openLogin = () => {
      loginService.openLogin();
    };

    return {
      authenticated,
      username,
      openLogin,
      store,
    };
  },
});
